+++ src/main/java/org/betonquest/betonquest/quest/condition/point/GlobalPointCondition.java
@@ -1,11 +1,14 @@
 package org.betonquest.betonquest.quest.condition.point;
 
+import org.betonquest.betonquest.Point;
 import org.betonquest.betonquest.api.profiles.Profile;
 import org.betonquest.betonquest.api.quest.condition.nullable.NullableCondition;
 import org.betonquest.betonquest.database.GlobalData;
 import org.betonquest.betonquest.exceptions.QuestRuntimeException;
 import org.betonquest.betonquest.instruction.variable.VariableNumber;
 import org.jetbrains.annotations.Nullable;
+
+import java.util.List;
 
 /**
  * A condition that checks if global data has a certain amount of points.
@@ -49,8 +52,17 @@ public class GlobalPointCondition implements NullableCondition {
 
     @Override
     public boolean check(@Nullable final Profile profile) throws QuestRuntimeException {
-        final int points = globalData.hasPointsFromCategory(category);
-        final int pCount = count.getValue(profile).intValue();
+        final List<Point> points = globalData.getPoints();
+        for (final Point point : points) {
+            if (point.getCategory().equals(category)) {
+                return checkPoints(point.getCount(), profile);
+            }
+        }
+        return false;
+    }
+
+    private boolean checkPoints(final int points, @Nullable final Profile profile) throws QuestRuntimeException {
+        final int pCount = this.count.getValue(profile).intValue();
         return equal ? points == pCount : points >= pCount;
     }
 }